resolv: avoid signed/unsigned comparisons

+++ src/config.c
@@ -600,7 +600,7 @@ accept_resolver_search(struct ResolverConfig *resolver, char *search) {
 
 static int
 accept_resolver_mode(struct ResolverConfig *resolver, char *mode) {
-    for (int i = 0; i < sizeof(resolver_mode_names) / sizeof(resolver_mode_names[0]); i++)
+    for (size_t i = 0; i < sizeof(resolver_mode_names) / sizeof(resolver_mode_names[0]); i++)
         if (strncasecmp(resolver_mode_names[i], mode, strlen(mode)) == 0) {
             resolver->mode = i;
             return 1;
